+++ src/apps/ScopeChangeRequest/hooks/sidesheet/useSidesheetEffects.tsx
@@ -34,7 +34,7 @@ export function useSidesheetEffects(
         if (!deref(sideSheetEditModeAtom)) {
             menuItems.push({
                 icon: <Icon name="edit" color={tokens.colors.interactive.primary__resting.hex} />,
-                label: 'Edit',
+                label: 'Edit request',
                 isDisabled: !canPatch,
                 onClick: toggleEditMode,
             });
@@ -43,7 +43,7 @@ export function useSidesheetEffects(
         menuItems.push(
             isVoided
                 ? {
-                    label: 'Unvoid',
+                    label: 'Unvoid request',
                     onClick: () => unVoidRequestMutation({ requestId }),
                     isDisabled: !canUnVoid,
                     icon: (
@@ -54,13 +54,13 @@ export function useSidesheetEffects(
                     ),
                 }
                 : {
-                    label: 'Void',
+                    label: 'Void request',
                     onClick: () => voidRequestMutation({ requestId }),
                     isDisabled: !canVoid,
                     icon: (
                         <Icon
                             name="delete_to_trash"
-                            color={tokens.colors.interactive.primary__resting.hex}
+                            color={tokens.colors.interactive.danger__resting.hex}
                         />
                     ),
                 }