+++ src/packages/BookmarksManager/src/components/BookmarksSidesheet/groupBookmarksBySubSystemAppKey.ts
@@ -1,0 +1,17 @@
+import { BookmarkResponse } from '../../types';
+
+type BookmarksBySubSystemAppKey = Record<string, Record<string, BookmarkResponse[]>>;
+export const groupBookmarksBySubSystemAppkey = (bookmarks: BookmarkResponse[]) => {
+    const bookmarksBySubSystem = bookmarks.reduce((acc, curr) => {
+        acc[curr.sourceSystem.subSystem] = {
+            ...(acc[curr.sourceSystem.subSystem] ? acc[curr.sourceSystem.subSystem] : {}),
+            [curr.appKey]: acc[curr.sourceSystem.subSystem]?.[curr.appKey]
+                ? [...acc[curr.sourceSystem.subSystem][curr.appKey], curr]
+                : [curr],
+        };
+
+        return acc;
+    }, {} as BookmarksBySubSystemAppKey);
+
+    return bookmarksBySubSystem;
+};